Soft-finalize classes that might need to be changed in the future
The classes changed in this commit inherit from Symfony interfaces. Changes in Symfony 8 will make it necessary to change interfaces in these classes in the near future.

By soft-finalizing these classes and making them final in the next major we gain legroom to make these changes without making it an API change visible to downstream users.

This is a squash-merge of https://github.com/FriendsOfBehat/MinkExtension/pull/41.

Squashed commit of the following:

commit f0a70e8e0fc8e60fe0a80e69e30f445bf0178df6
Author: Matthias Pigulla <mp@webfactory.de>
Date:   Wed Dec 17 11:29:57 2025 +0100

    Add a changelog entry

commit c58865c703424cb0f26515e325407a9faf91955c
Author: Matthias Pigulla <mp@webfactory.de>
Date:   Wed Dec 17 10:17:14 2025 +0100

    Soft-finalize classes that might need to be changed in the future