Mark Goutte driver/support as abandoned
Goutte has been abandoned, and depending on it blocks us from e. g. bumping dependencies or the CI environment to newer Symfony and PHP versions.

So, this announces the removal of Goutte support in the next major.

Unfortunately, Behat has not built in deprecation tracking system (yet). Anyways, we cannot defer such changes until it has.

This is a squash-merge of https://github.com/FriendsOfBehat/MinkExtension/pull/39/.

Squashed commit of the following:

commit 8fec8bc2ceee604f5addfd3ace931541bd1de472
Author: Matthias Pigulla <mp@webfactory.de>
Date:   Wed Dec 17 11:36:28 2025 +0100

    Add a changelog entry

commit 6a62541a3628b7a4e4020c8edb7cf9ab064dacc7
Author: Matthias Pigulla <mp@webfactory.de>
Date:   Wed Dec 17 09:50:31 2025 +0100

    Mark Goutte driver/support as abandoned